backend/pyproject.toml
[project]
name = "polkadot-relayer"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "fastapi>=0.135.1",
    "pydantic>=2.12.5",
    "python-dotenv>=1.2.2",
    "uvicorn>=0.42.0",
    "web3>=7.14.1",
]
